fix for #1240 - tried to remove all circular dependencies
