+++ packages/core/src/middleware.test.ts
@@ -4,7 +4,8 @@ import fetch from "isomorphic-unfetch"
 import {apiResolver} from "next/dist/next-server/server/api-utils"
 
 import {BlitzApiRequest, BlitzApiResponse} from "."
-import {Middleware, handleRequestWithMiddleware} from "./middleware"
+import {handleRequestWithMiddleware} from "./middleware"
+import {Middleware} from "./types"
 
 const testIfNotWindows = process.platform === "win32" ? it.skip : it
