+++ packages/core/src/middleware.ts
@@ -3,35 +3,16 @@ import {BlitzApiRequest, BlitzApiResponse} from "."
 import {IncomingMessage, ServerResponse} from "http"
 import {getConfig} from "@blitzjs/config"
 import {log, baseLogger} from "@blitzjs/display"
-import {EnhancedResolver} from "./types"
+import {
+  EnhancedResolver,
+  Middleware,
+  MiddlewareNext,
+  MiddlewareRequest,
+  MiddlewareResponse,
+} from "./types"
 
 export interface DefaultCtx {}
 export interface Ctx extends DefaultCtx {}
-
-export interface MiddlewareRequest extends BlitzApiRequest {
-  protocol?: string
-}
-export interface MiddlewareResponse extends BlitzApiResponse {
-  /**
-   * This will be passed as the second argument to Blitz queries/mutations.
-   *
-   * You must set blitzCtx BEFORE calling next()
-   */
-  blitzCtx: Record<string, unknown>
-  /**
-   * This is the exact result returned from the Blitz query/mutation
-   *
-   * You must first `await next()` before reading this
-   */
-  blitzResult: unknown
-}
-export type MiddlewareNext = (error?: Error) => Promise<void> | void
-
-export type Middleware = (
-  req: MiddlewareRequest,
-  res: MiddlewareResponse,
-  next: MiddlewareNext,
-) => Promise<void> | void
 
 export type ConnectMiddleware = (
   req: IncomingMessage,
