+++ packages/installer/src/recipe-builder.ts
@@ -1,7 +1,8 @@
-import {RecipeMeta, RecipeExecutor, ExecutorConfigUnion} from "./recipe-executor"
+import {RecipeExecutor, ExecutorConfigUnion} from "./recipe-executor"
 import * as AddDependencyExecutor from "./executors/add-dependency-executor"
 import * as NewFileExecutor from "./executors/new-file-executor"
 import * as TransformFileExecutor from "./executors/file-transform-executor"
+import {RecipeMeta} from "./types"
 
 export interface IRecipeBuilder {
   setName(name: string): IRecipeBuilder
