+++ packages/installer/src/types.ts
@@ -1,0 +1,6 @@
+export interface RecipeMeta {
+  name: string
+  description: string
+  owner: string
+  repoLink: string
+}
